+++ components/model/contents/index.ts
@@ -1,2 +1,1 @@
-export * from './SideBar';
 export * from './Contents';
