+++ components/page/Main.page.tsx
@@ -1,0 +1,7 @@
+import { MainContents } from '@/components/page/components/MainContents';
+
+const MainPage = () => {
+  return <MainContents />;
+};
+
+export default MainPage;
