+++ components/page/index.ts
@@ -1,0 +1,2 @@
+export * from './Main.page';
+export * from './components/MainContents';
