+++ src/pages/index.tsx
@@ -1,15 +1,3 @@
-import { Layout } from 'antd';
-import { NextPage } from 'next';
+import MainPage from '@/components/page/Main.page';
 
-import { Contents, SideBar } from '@/components/base/';
-
-const Home: NextPage = () => {
-  return (
-    <Layout className="bg-white min-h-40">
-      <SideBar />
-      <Contents />
-    </Layout>
-  );
-};
-
-export default Home;
+export default MainPage;
