+++ tsconfig.eslint.json
@@ -1,5 +1,11 @@
 {
   "extends": "./tsconfig.json",
-  "includes": ["src/**/*.ts", "src/**/*.tsx", ".eslintrc.json"],
+  "includes": [
+    "src/**/*.ts",
+    "src/**/*.tsx",
+    ".eslintrc.json",
+    "components/**/*.ts",
+    "components/**/*.tsx"
+  ],
   "exclude": ["node_modules", "dist"]
 }
